Change "default" wallet concept to "last active"
Now, when the user selects a wallet to be the active wallet, it will be
remembered the next time they sign in.

Some database procedures were changed to better reflect the renaming of
the concept (default to last active). But the underlying column name
(isDefault) in the database has not changed due to the difficulty in
renaming them properly.

Next steps: make sure the keystores are displayed on the portal in the
same order no matter which one is the active wallet.

Relates to DCP-577.