feat: update job-utility to use updated send() spec
The new function signature for send() specifies that only one
argument should be passed in, which is an instance of Message.
Previously, you could pass in arbitrary data which would then be
converted into a Message object for you.

This commit modifies the send call of the utility to pass in an
instance of Request instead of just a string.